fix: do not use statics so discover_peers can be called multiple times
Before this commit, scanned peers would be saved in a static Connector.
Which means that the scanned peers could not be forgotten without a restart of the program calling the lib.
This commit change a lot the backend by stopping using an AddressBook Service from cuprate but replicates some private functions to manage manually the saving of scanned nodes.